+++ com.unity.renderstreaming/Runtime/Scripts/InputSystem/Receiver.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 834a3bd628e331943b599618b7309d1f
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
