+++ com.unity.renderstreaming/Runtime/Scripts/InputSystem/Sender.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 1ba4fd50aea9ff3459e460141a546b3c
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
